corrigido erro do servidor deletar o proprio index.php

+++ server/curl.php
@@ -11,8 +11,11 @@ require 'crypto_rnd.php';
 //Limpa o diretorio de cookies
 //Deleta arquivos na pasta cookie que não foram modificados a mais de meia hora
 foreach (glob('cookies/'."*") as $file) {
-	if (filemtime($file) < time() - 1800) {
-		unlink($file);
+	if ($file != 'cookies/index.php')
+	{
+		if (filemtime($file) < time() - 1800) {
+			unlink($file);
+		}
 	}
 }
